REFACTOR : 마감시간 TimerTask 모듈화
