+++ app/src/main/kotlin/com/fakedevelopers/bidderbidder/ui/product_list/ProductListAdapter.kt
@@ -1,7 +1,6 @@
 package com.fakedevelopers.bidderbidder.ui.product_list
 
 import android.content.Context
-import android.os.CountDownTimer
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -15,9 +14,7 @@ import com.fakedevelopers.bidderbidder.api.data.Constants.Companion.dec
 import com.fakedevelopers.bidderbidder.databinding.RecyclerProductListBinding
 import com.fakedevelopers.bidderbidder.databinding.RecyclerProductListFooterBinding
 import com.fakedevelopers.bidderbidder.ui.product_list.ProductListViewModel.Companion.LIST_COUNT
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.TimeZone
+import com.fakedevelopers.bidderbidder.ui.util.ExpirationTimerTask
 
 class ProductListAdapter(
     private val clickLoadMore: () -> Unit,
@@ -26,29 +23,24 @@ class ProductListAdapter(
 ) : ListAdapter<ProductListDto, RecyclerView.ViewHolder>(diffUtil) {
 
     private var listSize = 0
-    private val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm").apply {
-        timeZone = TimeZone.getTimeZone("Asia/Seoul")
-    }
 
     inner class ItemViewHolder(
         private val binding: RecyclerProductListBinding,
         private val context: Context
     ) : RecyclerView.ViewHolder(binding.root) {
-        private lateinit var timerTask: CountDownTimer
+        private lateinit var timerTask: ExpirationTimerTask
         fun bind(item: ProductListDto) {
             with(binding) {
                 if (::timerTask.isInitialized) {
                     timerTask.cancel()
                 }
-                timerTask = object : CountDownTimer(getRemainTimeMillisecond(item.expirationDate), 1000) {
-                    override fun onTick(millisUntilFinished: Long) {
-                        textviewProductListExpire.text = getRemainTimeString(millisUntilFinished)
-                    }
-
-                    override fun onFinish() {
-                        textviewProductListExpire.text = "마감"
-                    }
-                }.start()
+                timerTask = ExpirationTimerTask(
+                    item.expirationDate,
+                    1000,
+                    tick = { remainTimeString -> textviewProductListExpire.text = remainTimeString },
+                    finish = { textviewProductListExpire.text = "마감" }
+                )
+                timerTask.start()
                 Glide.with(context)
                     .load(BASE_URL + item.thumbnail)
                     .placeholder(R.drawable.the_cat)
@@ -71,35 +63,6 @@ class ProductListAdapter(
                     clickProductDetail(item.productId)
                 }
             }
-        }
-
-        private fun getRemainTimeMillisecond(expirationDate: String) =
-            dateFormat.parse(expirationDate)!!.time - dateFormat.parse(dateFormat.format(Date()))!!.time
-
-        private fun getRemainTimeString(millisUntilFinished: Long): String {
-            val totalMinute = millisUntilFinished / 60000
-            val day = totalMinute / 1440
-            val hour = totalMinute % 1440 / 60
-            val remainTimeString = StringBuilder("마감까지 ")
-            // 일
-            if (day > 0) {
-                remainTimeString.append("${day}일 ")
-            }
-            // 시간
-            if (hour != 0L) {
-                remainTimeString.append("${hour}시간 ")
-            }
-            // 분, 초
-            if (day == 0L && hour < 3) {
-                val minute = totalMinute % 60
-                if (minute != 0L) {
-                    remainTimeString.append("${minute}분 ")
-                }
-                if (hour == 0L && minute < 5) {
-                    remainTimeString.append("${millisUntilFinished % 60000 / 1000}초")
-                }
-            }
-            return remainTimeString.toString()
         }
     }
